refactor(mcp-tools): 提取 MCP 工具通用工具函数至 utils 目录
