+++ src/renderer/pages/settings/tools/utils/index.ts
@@ -1,0 +1,7 @@
+export const truncateErrorMessage = (message: string, maxLength = 150): string => {
+  if (message.length <= maxLength) {
+    return message;
+  }
+
+  return `${message.substring(0, maxLength)}...`;
+};
